+++ src/lib.rs
@@ -9,20 +9,20 @@ doc_comment::doctest!("../README.md");
 #[cfg(test)]
 mod tests {
     use super::parser::Parser;
-    use super::types::{Event, HeadingLevel};
+    use super::types::{Event, HeadingLevel, Tag};
 
     #[test]
     fn parse_paragraphs() {
         let text = "abc\ndef\nghi\n\njkl";
         let mut parser = Parser::new(text);
-        assert_eq!(parser.next(), Some(Event::ParagraphBegin));
+        assert_eq!(parser.next(), Some(Event::Begin(Tag::Paragraph)));
         assert_eq!(parser.next(), Some(Event::Text("abc")));
         assert_eq!(parser.next(), Some(Event::Text("def")));
         assert_eq!(parser.next(), Some(Event::Text("ghi")));
-        assert_eq!(parser.next(), Some(Event::ParagraphEnd));
-        assert_eq!(parser.next(), Some(Event::ParagraphBegin));
+        assert_eq!(parser.next(), Some(Event::End(Tag::Paragraph)));
+        assert_eq!(parser.next(), Some(Event::Begin(Tag::Paragraph)));
         assert_eq!(parser.next(), Some(Event::Text("jkl")));
-        assert_eq!(parser.next(), Some(Event::ParagraphEnd));
+        assert_eq!(parser.next(), Some(Event::End(Tag::Paragraph)));
         assert_eq!(parser.next(), None);
     }
 
@@ -30,9 +30,15 @@ mod tests {
     fn parse_heading() {
         let text = "## abc";
         let mut parser = Parser::new(text);
-        assert_eq!(parser.next(), Some(Event::HeadingBegin(HeadingLevel::H2)));
+        assert_eq!(
+            parser.next(),
+            Some(Event::Begin(Tag::Heading(HeadingLevel::H2)))
+        );
         assert_eq!(parser.next(), Some(Event::Text("abc")));
-        assert_eq!(parser.next(), Some(Event::HeadingEnd(HeadingLevel::H2)));
+        assert_eq!(
+            parser.next(),
+            Some(Event::End(Tag::Heading(HeadingLevel::H2)))
+        );
         assert_eq!(parser.next(), None);
     }
 
@@ -40,9 +46,15 @@ mod tests {
     fn parse_heading_with_whitespaces() {
         let text = "##  abc";
         let mut parser = Parser::new(text);
-        assert_eq!(parser.next(), Some(Event::HeadingBegin(HeadingLevel::H2)));
+        assert_eq!(
+            parser.next(),
+            Some(Event::Begin(Tag::Heading(HeadingLevel::H2)))
+        );
         assert_eq!(parser.next(), Some(Event::Text("abc")));
-        assert_eq!(parser.next(), Some(Event::HeadingEnd(HeadingLevel::H2)));
+        assert_eq!(
+            parser.next(),
+            Some(Event::End(Tag::Heading(HeadingLevel::H2)))
+        );
         assert_eq!(parser.next(), None);
     }
 
@@ -50,9 +62,15 @@ mod tests {
     fn parse_heading_with_empty_text() {
         let text = "##";
         let mut parser = Parser::new(text);
-        assert_eq!(parser.next(), Some(Event::HeadingBegin(HeadingLevel::H2)));
+        assert_eq!(
+            parser.next(),
+            Some(Event::Begin(Tag::Heading(HeadingLevel::H2)))
+        );
         assert_eq!(parser.next(), Some(Event::Text("")));
-        assert_eq!(parser.next(), Some(Event::HeadingEnd(HeadingLevel::H2)));
+        assert_eq!(
+            parser.next(),
+            Some(Event::End(Tag::Heading(HeadingLevel::H2)))
+        );
         assert_eq!(parser.next(), None);
     }
 
@@ -60,9 +78,9 @@ mod tests {
     fn parse_heading_with_invalid_heading_level() {
         let text = "####### abc";
         let mut parser = Parser::new(text);
-        assert_eq!(parser.next(), Some(Event::ParagraphBegin));
+        assert_eq!(parser.next(), Some(Event::Begin(Tag::Paragraph)));
         assert_eq!(parser.next(), Some(Event::Text("####### abc")));
-        assert_eq!(parser.next(), Some(Event::ParagraphEnd));
+        assert_eq!(parser.next(), Some(Event::End(Tag::Paragraph)));
         assert_eq!(parser.next(), None);
     }
 }