+++ src/types/mod.rs
@@ -1,7 +1,9 @@
 mod block;
 mod event;
 mod heading_level;
+mod tag;
 
 pub use block::*;
 pub use event::*;
 pub use heading_level::*;
+pub use tag::*;