+++ src/types/tag.rs
@@ -1,0 +1,7 @@
+use crate::types::HeadingLevel;
+
+#[derive(Debug, PartialEq)]
+pub enum Tag {
+    Heading(HeadingLevel),
+    Paragraph,
+}